cleanup and refactor to use regular recoil
