+++ web/cypress/e2e/repository-details.cy.ts
@@ -55,7 +55,7 @@ describe('Repository Details Page', () => {
         'have.text',
         'See Child Manifests',
       );
-      cy.get(`[data-label="Size"]`).should('have.text', '2.51 kB - 4.12 kB');
+      cy.get(`[data-label="Size"]`).should('have.text', '2.51 kB ~ 4.12 kB');
       cy.get(`[data-label="Last Modified"]`).should(
         'have.text',
         formatDate('Thu, 04 Nov 2022 19:15:15 -0000'),
