+++ web/src/components/Table/ImageSize.tsx
@@ -1,40 +1,9 @@
 import {Skeleton} from '@patternfly/react-core';
-import {useEffect, useState} from 'react';
-import {
-  getManifestByDigest,
-  ManifestByDigestResponse,
-} from 'src/resources/TagResource';
 import prettyBytes from 'pretty-bytes';
-import {useRecoilState} from 'recoil';
+import {useEffect} from 'react';
+import {useSetRecoilState} from 'recoil';
 import {childManifestSizeState} from 'src/atoms/TagListState';
-
-export function useImageSize(org: string, repo: string, digest: string) {
-  const [size, setSize] = useState<number>(null);
-  const [loading, setLoading] = useState<boolean>(true);
-  const [err, setErr] = useState<boolean>(false);
-
-  useEffect(() => {
-    (async () => {
-      try {
-        const manifestResp: ManifestByDigestResponse =
-          await getManifestByDigest(org, repo, digest);
-        const calculatedSizeMesnifestResp = manifestResp.layers
-          ? manifestResp.layers.reduce(
-              (prev, curr) => prev + curr.compressed_size,
-              0,
-            )
-          : 0;
-        setSize(calculatedSizeMesnifestResp);
-      } catch (err) {
-        setErr(true);
-      } finally {
-        setLoading(false);
-      }
-    })();
-  }, [digest]);
-
-  return {size, loading, err};
-}
+import {useImageSize} from 'src/hooks/UseImageSize';
 
 export function ImageSize(props: ImageSizeProps) {
   const {size, loading, err} = useImageSize(
@@ -65,7 +34,7 @@ export function ChildManifestSize(props: ImageSizeProps) {
     props.digest,
   );
 
-  const [, setChildManifestSize] = useRecoilState(
+  const setChildManifestSize = useSetRecoilState(
     childManifestSizeState(props.digest),
   );
